add form_object

+++ app/models/cart.rb
@@ -1,10 +1,14 @@
 class Cart < ApplicationRecord
   has_many :line_items, dependent: :destroy
 
-  def add_product(product_id)
+  def add_product(product_id, quantity = nil)
     current_item = line_items.find_by(product_id: product_id)
     if current_item
-      current_item.quantity += 1
+      current_item.quantity += if quantity
+                                 quantity.to_i
+                               else
+                                 1
+                               end
     else
       current_item = line_items.build(product_id: product_id)
     end
@@ -13,16 +17,18 @@ class Cart < ApplicationRecord
 
   def del_product(product_id)
     current_item = line_items.find_by(product_id: product_id)
-    if current_item.quantity > 1
-      current_item.quantity -= 1
-    else
-      current_item.destroy
+    unless current_item.nil?
+      if current_item.quantity > 1
+        current_item.quantity -= 1
+      else
+        current_item.destroy
+      end
     end
     current_item
   end
 
   def destroy_product(product_id)
-    urrent_item = line_items.find_by(product_id: product_id)
+    current_item = line_items.find_by(product_id: product_id)
     current_item.destroy
   end
